# src/main/resources/static/swagger.yaml
swagger: '2.0'
info:
  description: This page lists all the rest apis for Demo Swagger App.
  version: 1.0-SNAPSHOT
  title: Demo Swagger Rest APIs
host: 'localhost:8080'
basePath: /
tags:
  - name: catalog-controller
    description: Catalog Controller
paths:
  '/delete/{id}':
    delete:
      tags:
        - catalog-controller
      summary: deleteItem
      operationId: deleteItemUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            type: object
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /getProduct:
    get:
      tags:
        - catalog-controller
      summary: catalog
      operationId: catalogUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/getProduct/{id}':
    get:
      tags:
        - catalog-controller
      summary: catalogById
      operationId: catalogByIdUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /save:
    post:
      tags:
        - catalog-controller
      summary: create
      operationId: createUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: json
          description: json
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          schema:
            type: integer
            format: int32
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  Product:
    type: object
    properties:
      description:
        type: string
      id:
        type: integer
        format: int32
      instock:
        type: integer
        format: int32
      name:
        type: string
      price:
        type: integer
        format: int32
      warranty:
        type: integer
        format: int32